requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
python-dotenv==1.0.1
werkzeug==3.1.3
jinja2==3.1.6
sqlalchemy==2.0.39
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
wtforms==3.2.1
pywin32==309
plotly==5.18.0
pandas==2.2.0
numpy==1.26.4 